uxperf workloads: only broadcast media mounted if needed
Previously AndroidPerfWorkload broadcast media mounted unconditionally
inside push_assets and delete_assets. This change makes it so the
broadcast only happens if something was actually pushed/needs to be
delete.

+++ wlauto/common/android/workload.py
@@ -652,17 +652,21 @@ class AndroidUxPerfWorkload(AndroidUiAutoBenchmark):
         return self.device.path.join(dirname, fname)
 
     def push_assets(self, context):
+        pushed = False
         for f in self.deployable_assets:
             fpath = context.resolver.get(File(self, f))
             device_path = self._path_on_device(fpath)
             if self.force_push_assets or not self.device.file_exists(device_path):
                 self.device.push_file(fpath, device_path, timeout=300)
-        self.device.broadcast_media_mounted(self.device.working_directory)
+                pushed = True
+        if pushed:
+            self.device.broadcast_media_mounted(self.device.working_directory)
 
     def delete_assets(self):
-        for f in self.deployable_assets:
-            self.device.delete_file(self._path_on_device(f))
-        self.device.broadcast_media_mounted(self.device.working_directory)
+        if self.deployable_assets:
+            for f in self.deployable_assets:
+                self.device.delete_file(self._path_on_device(f))
+            self.device.broadcast_media_mounted(self.device.working_directory)
 
     def __init__(self, device, **kwargs):
         super(AndroidUxPerfWorkload, self).__init__(device, **kwargs)
